+++ Package.swift
@@ -6,9 +6,11 @@ let package = Package(
     platforms: [.iOS(.v15), .macOS(.v12)],
     products: [
         .library(name: "ATCParserKit", targets: ["ATCParserKit"]),
+        .library(name: "ATCParserFFI", targets: ["ATCParserFFI"]),
     ],
     targets: [
         .target(name: "ATCParserKit"),
+        .target(name: "ATCParserFFI", dependencies: ["ATCParserKit"]),
         .testTarget(name: "ATCParserKitTests", dependencies: ["ATCParserKit"]),
     ]
 )